[rename] rag_service -> chat_service로 변경
